requirements.txt
duolingo-api==0.5.3
fbchat==1.9.7